Updated ClientDB structures to 7.0.1.20756

+++ World of Warcraft/Bins/DataExtractor/Structures/GameTables.cs
@@ -6,7 +6,7 @@ namespace DataExtractor.Structures
     {
         public int Index;
         public string Name;
-        public int NumRows;
-        public int NumColumns;
+        public short NumRows;
+        public byte NumColumns;
     }
 }
